+++ x/distribution/legacy/v043/helpers.go
@@ -1,4 +1,4 @@
-package v042
+package v043
 
 import (
 	"github.com/cosmos/cosmos-sdk/store/prefix"
