+++ x/distribution/legacy/v043/store.go
@@ -1,4 +1,4 @@
-package v042
+package v043
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
