+++ x/slashing/keeper/migrations.go
@@ -2,7 +2,7 @@ package keeper
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	v042 "github.com/cosmos/cosmos-sdk/x/slashing/legacy/v042"
+	v043 "github.com/cosmos/cosmos-sdk/x/slashing/legacy/v043"
 )
 
 // Migrator is a struct for handling in-place store migrations.
@@ -17,5 +17,5 @@ func NewMigrator(keeper Keeper) Migrator {
 
 // Migrate1to2 migrates from version 1 to 2.
 func (m Migrator) Migrate1to2(ctx sdk.Context) error {
-	return v042.MigrateStore(ctx, m.keeper.storeKey)
+	return v043.MigrateStore(ctx, m.keeper.storeKey)
 }
