+++ x/slashing/legacy/v043/store.go
@@ -1,8 +1,8 @@
-package v042
+package v043
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	v042distribution "github.com/cosmos/cosmos-sdk/x/distribution/legacy/v042"
+	v043distribution "github.com/cosmos/cosmos-sdk/x/distribution/legacy/v043"
 	v040slashing "github.com/cosmos/cosmos-sdk/x/slashing/legacy/v040"
 )
 
@@ -12,9 +12,9 @@ import (
 // - Change addresses to be length-prefixed.
 func MigrateStore(ctx sdk.Context, storeKey sdk.StoreKey) error {
 	store := ctx.KVStore(storeKey)
-	v042distribution.MigratePrefixAddress(store, v040slashing.ValidatorSigningInfoKeyPrefix)
-	v042distribution.MigratePrefixAddressBytes(store, v040slashing.ValidatorMissedBlockBitArrayKeyPrefix)
-	v042distribution.MigratePrefixAddress(store, v040slashing.AddrPubkeyRelationKeyPrefix)
+	v043distribution.MigratePrefixAddress(store, v040slashing.ValidatorSigningInfoKeyPrefix)
+	v043distribution.MigratePrefixAddressBytes(store, v040slashing.ValidatorMissedBlockBitArrayKeyPrefix)
+	v043distribution.MigratePrefixAddress(store, v040slashing.AddrPubkeyRelationKeyPrefix)
 
 	return nil
 }
