+++ x/slashing/legacy/v043/store_test.go
@@ -1,4 +1,4 @@
-package v042_test
+package v043_test
 
 import (
 	"bytes"
@@ -10,7 +10,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/testutil/testdata"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	v040slashing "github.com/cosmos/cosmos-sdk/x/slashing/legacy/v040"
-	v042slashing "github.com/cosmos/cosmos-sdk/x/slashing/legacy/v042"
+	v043slashing "github.com/cosmos/cosmos-sdk/x/slashing/legacy/v043"
 	"github.com/cosmos/cosmos-sdk/x/slashing/types"
 )
 
@@ -52,7 +52,7 @@ func TestStoreMigration(t *testing.T) {
 	}
 
 	// Run migrations.
-	err := v042slashing.MigrateStore(ctx, slashingKey)
+	err := v043slashing.MigrateStore(ctx, slashingKey)
 	require.NoError(t, err)
 
 	// Make sure the new keys are set and old keys are deleted.
